pyproject.toml
[project]
name = "emr-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "celery>=5.6.3",
    "fastapi>=0.135.3",
    "openai>=2.30.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
    "greenlet>=3.1.1",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.24",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.33.1",
]
